LayoutTests/platform/win/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-willValidate-expected.txt



PASS [INPUT in HIDDEN status] Must be barred from the constraint validation
PASS [INPUT in BUTTON status] Must be barred from the constraint validation
PASS [INPUT in RESET status] Must be barred from the constraint validation
PASS [BUTTON in BUTTON status] Must be barred from the constraint validation
PASS [BUTTON in RESET status] Must be barred from the constraint validation
FAIL [fieldset]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false
FAIL [output]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false
PASS [object]  Must be barred from the constraint validation
PASS [keygen]  Must be barred from the constraint validation
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in TEXT status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in TEXT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in SEARCH status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in SEARCH status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in TEL status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in TEL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in URL status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in URL status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in URL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in URL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in EMAIL status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in EMAIL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in PASSWORD status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in PASSWORD status] The willValidate attribute must be false if it has a datalist ancestor
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"
FAIL [INPUT in DATE status] The date type must be supported. assert_equals: The date type should be supported. expected "date" but got "text"
FAIL [INPUT in MONTH status] The month type must be supported. assert_equals: The month type should be supported. expected "month" but got "text"
FAIL [INPUT in WEEK status] The week type must be supported. assert_equals: The week type should be supported. expected "week" but got "text"
FAIL [INPUT in TIME status] The time type must be supported. assert_equals: The time type should be supported. expected "time" but got "text"
FAIL [INPUT in COLOR status] The color type must be supported. assert_equals: The color type should be supported. expected "color" but got "text"
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in FILE status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in FILE status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is disabled
FAIL [INPUT in SUBMIT status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [BUTTON in SUBMIT status] Must be barred from the constraint validation
FAIL [BUTTON in SUBMIT status] The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [BUTTON in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [select]  Must be barred from the constraint validation
FAIL [select]  The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [select]  The willValidate attribute must be false if it has a datalist ancestor
PASS [textarea]  Must be barred from the constraint validation
FAIL [textarea]  The willValidate attribute must be true if an elment is mutable assert_true: The willValidate attribute should be true. expected true got false
PASS [textarea]  The willValidate attribute must be false if it has a datalist ancestor
